light refact
